Full validated multi contrast multi session unit tests with updated reference files
